feat(suitesparse-dl): download matrices into different NNZ categories
we can divide the matrices into different categories by number of non-zeros, and save files to
different category directories.

+++ tools/suitesparse-dl/mtx/matrix.go
@@ -5,13 +5,15 @@ type DlLinks struct {
 	RutherfordBoeing string // Rutherford Boeing format
 	MatrixMarket     string // Matrix Market format
 }
+
+// MatrixMeta described the base information of a matrix
 type MatrixMeta struct {
 	ID      string
 	Name    string
 	Group   string
-	Rows    int
-	Cols    int
-	NNZ     int
+	Rows    int64
+	Cols    int64
+	NNZ     int64 // we use int64, nnz can be very large (can be larger than 2^32)
 	Kind    string
 	Date    string
 	DlLinks DlLinks
